+++ module/Core/src/ShortUrl/Model/Validation/DeviceLongUrlsValidator.php
@@ -11,9 +11,9 @@ use Shlinkio\Shlink\Core\Model\DeviceType;
 use function array_keys;
 use function array_values;
 use function is_array;
-use function Shlinkio\Shlink\Core\contains;
+use function Shlinkio\Shlink\Core\ArrayUtils\contains;
+use function Shlinkio\Shlink\Core\ArrayUtils\every;
 use function Shlinkio\Shlink\Core\enumValues;
-use function Shlinkio\Shlink\Core\every;
 
 class DeviceLongUrlsValidator extends AbstractValidator
 {