+++ module/Core/test/Importer/ImportedLinksProcessorTest.php
@@ -32,8 +32,8 @@ use stdClass;
 use Symfony\Component\Console\Style\StyleInterface;
 
 use function count;
-use function Shlinkio\Shlink\Core\contains;
-use function Shlinkio\Shlink\Core\some;
+use function Shlinkio\Shlink\Core\ArrayUtils\contains;
+use function Shlinkio\Shlink\Core\ArrayUtils\some;
 use function sprintf;
 use function str_contains;